samples/python/requirements.txt
PyYAML==6.0
jmespath==1.0.1
